refactor: Number클래스 별개의 파일로 추출
